probably fixed a pYtorch error

+++ bert.py
@@ -45,8 +45,8 @@ def main():
     attention_mask = np.where(padded != 0, 1, 0)
     attention_mask.shape
 
-    input_ids = torch.tensor(padded)
-    attention_mask = torch.tensor(attention_mask)
+    input_ids = torch.tensor(padded).to(torch.int64)
+    attention_mask = torch.tensor(attention_mask).to(torch.int64)
 
     with torch.no_grad():
         last_hidden_states = model(input_ids, attention_mask=attention_mask)
